# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the HopSpot bench management app",
        "title": "HopSpot API",
        "contact": {
            "name": "Christoph Knuchel",
            "email": "christoph.knuchel@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/benches/{id}/visits/count": {
            "get": {
                "description": "Retrieve the total number of visits for a specific bench",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Visits"
                ],
                "summary": "Get visit count by bench ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bench ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.VisitCountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/api/v1/visits": {
            "get": {
                "description": "Get a paginated list of visits for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Visits"
                ],
                "summary": "List visits",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.PaginatedVisitsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "description": "Record a new visit to a bench",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Visits"
                ],
                "summary": "Create a new visit",
                "parameters": [
                    {
                        "description": "Visit payload",
                        "name": "visit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateVisitRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.VisitResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.CreateVisitRequest": {
            "type": "object",
            "required": [
                "bench_id"
            ],
            "properties": {
                "bench_id": {
                    "type": "integer"
                },
                "comment": {
                    "type": "string",
                    "maxLength": 500
                },
                "visited_at": {
                    "type": "string",
                    "default": "now"
                }
            }
        },
...
        "responses.PaginatedVisitsResponse": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/responses.PaginationResponse"
                },
                "visits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.VisitResponse"
                    }
                }
            }
        },
        "responses.PaginationResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "responses.VisitBenchResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "responses.VisitCountResponse": {
            "type": "object",
            "properties": {
                "bench_id": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "responses.VisitResponse": {
            "type": "object",
            "properties": {
                "bench": {
                    "$ref": "#/definitions/responses.VisitBenchResponse"
                },
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "visited_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}